requirements.txt
fastapi
uvicorn
openai-whisper
torch
setuptools-rust
python-multipart
